% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cameraPR.matrix.R
\name{cameraPR.matrix}
\alias{cameraPR.matrix}
\title{Competitive Gene Set Test Accounting for Inter-gene Correlation}
\usage{
\method{cameraPR}{matrix}(
  statistic,
  index,
  use.ranks = FALSE,
  inter.gene.cor = 0.01,
  sort = TRUE,
  adjust.globally = FALSE
)
}
\arguments{
\item{statistic}{a matrix of statistics (e.g., moderated t-statistics;
possibly with \code{\link[base]{NA}}s) with genes as row names and one or
more contrasts as column names.}

\item{index}{a named list of sets to test. Passed to \code{\link{incidence}}.
\code{index} must be a list of genes (character vectors), not the result of
\code{\link[limma]{ids2indices}}, so it is more restrictive than what
\code{\link[limma]{cameraPR.default}} allows.}

\item{use.ranks}{logical; whether to perform a rank-based test (\code{TRUE})
or a parametric test (\code{FALSE}; default).}

\item{inter.gene.cor}{numeric; the inter-gene correlation within tested sets.
May be a single value or a named vector with names matching those of
\code{index}.}

\item{sort}{logical; should the results of each contrast be sorted by
p-value? Default is \code{TRUE}.}

\item{adjust.globally}{logical; whether p-values from different contrasts
should be adjusted together. It is recommended to set this to \code{TRUE}
when testing a set of closely related contrasts. See Section 13.3 of the
LIMMA User's Guide (\code{\link[limma]{limmaUsersGuide}}) for details.
Default is \code{FALSE}.}
}
\value{
A \code{data.frame} with the following columns:

\item{Contrast}{factor; the contrast of interest.}
\item{GeneSet}{character; the gene set being tested.}
\item{NGenes}{integer; number of genes in the set that were present in the
\code{statistic} matrix.}
\item{Correlation}{numeric; inter-gene correlation
(only included if \code{inter.gene.cor} was not a single value).}
\item{Direction}{character; direction of change ("Up" or "Down").}
\item{Statistic}{numeric; two-sample t-statistic (\code{use.ranks=FALSE})
or standard normal z-statistic.}
\item{df}{numeric; degrees of freedom (only included if
\code{use.ranks=FALSE}). Two less than the number of non-missing values in
each column of the \code{statistic} matrix.}
\item{PValue}{numeric; two-tailed p-value.}
\item{FDR}{numeric; Benjamini and Hochberg FDR adjusted p-value (not
included when only one set and contrast was tested).}
}
\description{
Pre-ranked Correlation-Adjusted MEan RAnk gene set testing
(CAMERA-PR) tests whether a set of genes is highly ranked relative to other
genes in terms of differential expression, accounting for inter-gene
correlation (Wu & Smyth, 2012). See \code{\link[camera]{cameraPR}} for
details.

While the language is centered around gene sets, any \emph{a priori} groups
of biomolecules may be tested.
}
\section{Test Assumptions}{


CAMERA and CAMERA-PR extend the two-sample t-test or Wilcoxon-Mann-Whitney
test (if \code{use.ranks=TRUE}) to allow for correlation of genes within
each set of interest, though genes not in a given set are assumed to be
independent (Wu & Smyth, 2012).

If \code{use.ranks=FALSE}, it is assumed that the statistics in each group
are sampled from Normal distributions with equal variance, though the
two-sample t-test is fairly robust to departures from normality when group
sizes are large (i.e., when the gene sets are not too small or too large
relative to the total number of non-missing values in each column of
\code{statistic}).

If \code{use.ranks=TRUE}, the ranks of the statistics will be used, so
there are no distributional assumptions.
}

\references{
Wu, D., and Smyth, G. K. (2012). Camera: a competitive gene set
test accounting for inter-gene correlation. \emph{Nucleic Acids Research}
40, e133.
\href{https://doi.org/10.1093/nar/gks461}{doi:10.1093/nar/gks461}.

Goeman, J. J., and Bühlmann, P. (2007). Analyzing gene expression data in
terms of gene sets: methodological issues. \emph{Bioinformatics} 23,
980-987.
\href{https://doi.org/10.1093/bioinformatics/btm051}{doi:10.1093/bioinformatics/btm051}.
}
\seealso{
\code\link{\link{limma}{cameraPR}},
\code{\link[limma]{rankSumTestWithCorrelation}}
}
\author{
Di Wu, Gordon Smyth, and Tyler Sagendorf
}
