% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_sets.R
\name{decompose_sets}
\alias{decompose_sets}
\title{Decompose pairs of overlapping sets into 3 disjoint parts}
\source{
Decomposition of sets is described by Jiang and Gentleman (2007) in
section 2.3.1 "Overlap among gene sets". It is a method to reduce the
redundancy of significant gene set testing results whereby the decomposed
sets are reanalyzed and the following selections can be made:

\itemize{
\item{If the elements unique to set 1 and set 2, elements common to both
sets, or all 3 parts are statistically significant, keep both set 1 and set
2 in the original results. We can not separate their effects.}

\item{If the elements unique to set 1 or the elements unique to set 1 and
common to both sets are statistically significant, only keep set 1 in the
original results. (The same logic can be applied for set 2.)}
}
}
\usage{
decompose_sets(x, overlap = 1L)
}
\arguments{
\item{x}{a named list of sets.}

\item{overlap}{integer; only pairs of sets with at least
\code{overlap} elements in common will be decomposed.}
}
\value{
A list of disjoint parts of sets. May contain aliases.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Decompose all pairs of sufficiently overlapping sets into 3 disjoint parts:
the elements unique to the first set, the elements unique to the second
set, and the elements found in both sets.
}
\section{Optimization}{


Since the size of the intersection between two sets is at most the size of
the smaller set, any sets with fewer than \code{overlap} elements can be
immediately discarded.
}

\examples{
x <- list("A" = letters[1:10],
          "B" = letters[3:7],
          "C" = letters[1:4],
          "D" = letters[6:12])

decompose_sets(x)

decompose_sets(x, overlap = 5L)
}
\references{
Jiang, Z., & Gentleman, R. (2007). Extensions to gene set enrichment.
\emph{Bioinformatics, 23}(3), 306–313.
\url{https://doi.org/10.1093/bioinformatics/btl599}
}
\seealso{
\code{\link{filter_sets}}
}
