% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmap_color_functions.R
\name{enrichmap_color_functions}
\alias{enrichmap_color_functions}
\alias{gsea_color_fun}
\alias{camera_color_fun}
\title{Color functions for enrichmap}
\usage{
gsea_color_fun(statistics, colors = c("#3366ff", "darkred"))

camera_color_fun(statistics, colors = c("#3366ff", "darkred"))
}
\arguments{
\item{statistics}{numeric matrix or vector of statistics. Missing values are
removed. Used to compute limits for the color legend.}

\item{colors}{vector of 2 colors for the most negative and most positive
values of \code{statistics}. Default is "#3366ff" (blue) and "darkred".}
}
\value{
a named list of breaks and colors for the heatmap legend.
}
\description{
Heatmap color functions for plotting GSEA-like and CAMERA-like
results with \code{\link{enrichmap}}.
}
\details{
For \code{gsea_color_fun}, the \code{statistics} are expected to be
normalized enrichment scores (NES). Due to how the NES is formulated,
values between -1 and 1 are never significant or otherwise interesting, so
they are given a white fill so as to not appear in the heatmap.
}
\seealso{
\code{\link{range_extend}}
}
