% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sets.R
\name{filter_sets}
\alias{filter_sets}
\title{Filter a Named List of Sets by Size}
\usage{
filter_sets(x, background, min_size = 5L, max_size = Inf)
}
\arguments{
\item{x}{a named list of sets. Elements must be of type \code{"character"}.}

\item{background}{character; optional character vector. \code{x} will be
filtered to only those elements of \code{background}. If \code{background}
is any other atomic vector, it will be coerced to type \code{character}.}

\item{min_size}{integer (\eqn{\geq 1}); the minimum allowable set size.}

\item{max_size}{integer (\eqn{\geq 1}); the maximum allowable set size.}
}
\value{
A named list of sets at most the same size as \code{x}.
}
\description{
Given a named list of sets, filter to those that contain at
least \code{min_size} and no more than \code{max_size} elements. The sets
are optionally restricted to elements of \code{background} before filtering
by size.
}
\examples{
x <- list("A" = c("a", "b", "c"),
          "B" = c("a", "a", "d", "e", NA), # duplicates and NA
          "C" = c("f", "g"))

# All sets have at least 2 elements, so this just removes duplicates and NA
filter_sets(x, min_size = 2L)

# Limit scope of sets before filtering
filter_sets(x, min_size = 2L, background = c("a", "c", "e", "z"))
}
