% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_aliases.R
\name{find_aliases}
\alias{find_aliases}
\title{List aliased sets}
\usage{
find_aliases(x)
}
\arguments{
\item{x}{a named list of sets.}
}
\value{
A list of aliases with primary sets as names. If no sets are
  aliased, a warning will be issued and the function will return a list with
  the same length as \code{x}, but all elements will be \code{character(0)}.
}
\description{
Given a list of sets, collapse groups of aliased sets to the
  first set by alphanumeric order and list the rest as aliases.
}
\section{Function Details}{


  \code{find_aliases} uses the fact that a Jaccard similarity coefficient of
  1 indicates that sets are aliased. Groups of aliased sets are identified
  with \code{\link{cluster_sets}}. For each group, the first set
  alphanumerically is taken as the set to keep, while all others are listed
  as aliases. All sets are pre-filtered with \code{\link{filter_sets}} to
  have at least 2 unique elements.
}

\examples{
## B, C, and D are aliased
x <- list("A" = letters[1:5],
          "B" = letters[1:4],
          "C" = letters[1:4],
          "D" = letters[1:4])

# Names are sets to keep. List elements are aliases
(aliases <- find_aliases(x))

# Remove aliases from x
(x <- x[names(aliases)])

## No sets are aliased
y <- list("A" = letters[1:3],
          "B" = letters[2:3])

find_aliases(y)

}
\seealso{
\code{\link{filter_sets}}, \code{\link{similarity}},
  \code{\link{cluster_sets}}
}
