% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt_to_list.R
\name{gmt_to_list}
\alias{gmt_to_list}
\title{Read GMT File to a Named List of Sets}
\usage{
gmt_to_list(path, check = TRUE)
}
\arguments{
\item{path}{character; path to a GMT file. Files may include one additional
extension after ".gmt", such as ".gmt.gzip".}

\item{check}{logical; check that \code{path} points to a valid GMT file. If
\code{FALSE}, files with different extensions may be read, so long as they
are in the expected format.}
}
\value{
A named list of character vectors.
}
\description{
Create a named list of sets from a GMT file, or a file
structured like a GMT.
}
\note{
Similar to \code{fgsea::gmtPathways}.
}
\examples{
path <- system.file("extdata", "c5.go.v2023.2.Hs.symbols.gmt.gz",
                    package = "most")

x <- gmt_to_list(path)

head(names(x)) # First 6 gene set names

x[1] # first set
}
