% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence}
\alias{incidence}
\title{Construct sparse incidence matrix from list of sets}
\usage{
incidence(x)
}
\arguments{
\item{x}{a named list of sets. Elements must be of type \code{"character"}.}
}
\value{
An object of class \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}
with unique set names as rows and unique elements as columns. A value of 1
indicates the element is a member of the set.
}
\description{
Given a named list of sets, construct a sparse incidence matrix.
}
\examples{
x <- list("A" = c("a", "b", "c"),
          "B" = c("c", "d"),
          "C" = c("x", "y", "z", "z"), # duplicates
          "D" = c("a", NA)) # missing values

(mat <- incidence(x))

# Sizes of sets and their intersections
tcrossprod(mat)

# Number of elements unique to each set
keep <- apply(mat, 2, sum) == 1
apply(mat[, keep], 1, sum)
}
\seealso{
\code{\link{incidence_to_list}}, \code{\link{similarity}},
\code{\link[Matrix]{sparseMatrix}}
}
