% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_sets.R
\name{invert_sets}
\alias{invert_sets}
\title{Invert a List of Sets, Transposing Sets and Elements}
\usage{
invert_sets(x)
}
\arguments{
\item{x}{a named list of sets. Elements must be of type \code{"character"}.}
}
\value{
A named list of sets.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Invert a list of sets so that elements become set names and set names
become elements.
}
\note{
This function is essentially a more limited version of
\code{purrr::transpose_list}.
}
\examples{
x <- list("A" = c("a", "b", "c"),
          "B" = c("c", "d"),
          "C" = c("x", "y", "z"),
          "D" = c("a", "c", "d"))

# Invert sets
(y <- invert_sets(x))

# Decompose sets into disjoint parts
x_dc <- lapply(y, paste, collapse = ", ")
(x_dc <- invert_sets(x_dc))
}
