% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_aliases.R
\name{remove_aliases}
\alias{remove_aliases}
\title{Remove aliased sets}
\usage{
remove_aliases(x)
}
\arguments{
\item{x}{a named list of sets.}
}
\value{
a named list of sets with aliases removed.
}
\description{
Remove aliased sets
}
\section{Function Details}{


  \code{remove_aliases} is a wrapper for \code{find_aliases}, which itself
  uses \code{cluster_sets}. It is recommended to use \code{find_aliases}
  first. This way, aliases can be identified and included in the final
  analysis results.

  All sets are pre-filtered to have at least 2 unique elements. For each
  group of aliased sets, the first set is kept according to alphanumeric
  order. All non-aliased sets are also kept.
}

\examples{
# A and B are aliased, C and D are aliased
x <- list("C" = 4:6,
          "B" = 1:3,
          "A" = 1:3,
          "D" = 4:6)

remove_aliases(x) # B and D removed
}
\seealso{
\code{\link{find_aliases}}, \code{\link{filter_sets}}
}
