% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMSig-package.R
\docType{package}
\name{TMSig-package}
\alias{TMSig}
\alias{TMSig-package}
\title{TMSig: Tools for Molecular Signatures}
\description{
The TMSig package contains tools to prepare, analyze, and visualize named lists of sets, with an emphasis on molecular signatures (such as gene or kinase sets). It includes fast, memory efficient functions to construct sparse incidence and similarity matrices and filter, cluster, invert, and decompose sets. Additionally, bubble heatmaps can be created to visualize the results of any differential or molecular signatures analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EMSL-Computing/TMSig}
  \item Report bugs at \url{https://github.com/EMSL-Computing/TMSig/issues}
}

}
\author{
\strong{Maintainer}: Tyler Sagendorf \email{tyler.sagendorf@pnnl.gov} (\href{https://orcid.org/0000-0003-1552-4870}{ORCID})

Other contributors:
\itemize{
  \item Di Wu [contributor]
  \item Gordon Smyth [contributor]
}

}
\keyword{internal}
