% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmap.R
\name{enrichmap}
\alias{enrichmap}
\title{Set Enrichment Bubble Heatmap}
\usage{
enrichmap(
  x,
  n_top = 15L,
  set_column = "GeneSet",
  statistic_column = "ZScore",
  contrast_column = "Contrast",
  padj_column = "PAdj",
  padj_legend_title = padj_column,
  padj_aggregate_fun = function(padj) {
     median(-log10(padj), na.rm = TRUE)
 },
  padj_cutoff = 0.05,
  plot_sig_only = TRUE,
  padj_fill = "grey45",
  colors = c("#3366ff", "darkred"),
  heatmap_color_fun = camera_color_fun,
  scale_by = c("row", "column", "max"),
  cell_size = unit(14, "points"),
  filename,
  height = 5,
  width = 5,
  units = "in",
  heatmap_args = list(),
  padj_args = list(),
  save_args = list(),
  draw_args = list()
)
}
\arguments{
\item{x}{an object that can be coerced to a \code{data.table} with columns
\code{contrast_column}, \code{set_column}, \code{statistic_column}, and
\code{padj_column}.}

\item{n_top}{integer; number of sets (rows) to display. Defaults to the top
15 sets with the highest median \eqn{-log_{10}}(adjusted p-values) across
contrasts.}

\item{set_column}{character; the name of a column in \code{x} containing
unique set identifiers that will be used as the row names in the heatmap.
Default is "GeneSet".}

\item{statistic_column}{similar to \code{set_column}. The name of a column
containing the statistic for each pathway. Determines the heatmap body
colors. Normalized enrichment scores (NES) are assumed, by default, so
values between -1 and +1 (noise) will appear white. This can be modified
through the use of \code{heatmap_args}.}

\item{contrast_column}{character; the name of a column in \code{x} containing
contrasts that will be used as columns for the heatmap. Entries of
\code{x[[rownames_colum]]} must be uniquely defined for each contrast
group.}

\item{padj_column}{similar to \code{set_column}. The name of a column
containing the adjusted p-values that determine the area of each circle in
the heatmap.}

\item{padj_legend_title}{character; title of the background fill legend.
Defaults to \code{padj_column}.}

\item{padj_aggregate_fun}{function; a function used to aggregate the adjusted
p-values in \code{x[[pvalue_column]]} across contrasts for each unique
entry in \code{x[[set_column]]}. Passed to \code{\link[base]{match.fun}}.
The default computes the median of the \eqn{-log_{10}} adjusted p-values.}

\item{padj_cutoff}{numeric; cutoff for terms to be statistically significant.
If \code{plot_sig_only=TRUE}, only those pathways with at least one
\code{padj_column} value less than this threshold may appear in the
heatmap. Default is 0.05.}

\item{plot_sig_only}{logical; whether to plot only those \code{n_top} terms
that have at least one \code{padj_column} value less than
\code{padj_cutoff}.}

\item{padj_fill}{character; the background color used for values in
\code{padj_column} that are less than \code{padj_cutoff}. Default is
"grey45".}

\item{colors}{vector of length 2 specifying the colors for the largest
negative and largest positive values of \code{x[[statistic_column]]},
respectively. Default is "#3366ff" (blue) and "darkred".}

\item{heatmap_color_fun}{function; used to create the legend for the heatmap
bubble fill. See \code{\link{enrichmap_color_functions}} for details.}

\item{scale_by}{character; whether to scale the circles such that the
most-significant term in each row (\code{scale_by="row"}), column
(\code{scale_by="column"}), or overall (\code{scale_by="max"}) is of
maximum area. Default is "row" to better visualize patterns across
contrasts.}

\item{cell_size}{\code{\link[grid]{unit}} object; the size of each heatmap
cell (used for both height and width). Default is \code{unit(14,
  "points")}.}

\item{filename}{character; the filename used to save the heatmap. If missing
(default), the heatmap will be displayed instead.}

\item{height}{numeric; height of the file in \code{units}.}

\item{width}{similar to \code{height}.}

\item{units}{character; units that define \code{height} and \code{width}.
Defaults to "in" (inches).}

\item{heatmap_args}{list; additional arguments passed to
\code{\link[ComplexHeatmap]{Heatmap}}.}

\item{padj_args}{list; additional arguments passed to
\code{\link[ComplexHeatmap]{Legend}}. Modifies the adjusted p-value legend.}

\item{save_args}{list; additional arguments passed to the graphics device
determined by the \code{filename} extension. See
\code{\link[grDevices]{png}} and \code{\link[grDevices]{pdf}} for options.}

\item{draw_args}{list; additional arguments passed to
\code{\link[ComplexHeatmap]{draw-HeatmapList-method}}.}
}
\value{
Nothing. Displays heatmap in Plots pane or saves the heatmap to a
file (if \code{filename} is provided).
}
\description{
Create a bubble heatmap summarizing set enrichment analysis
results.
}
\seealso{
\code{\link[ComplexHeatmap]{ComplexHeatmap-package}}
}
