% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sets.R
\name{filter_sets}
\alias{filter_sets}
\title{Filter a list of sets by size}
\usage{
filter_sets(x, background, min_size = 5L, max_size = Inf)
}
\arguments{
\item{x}{a named list of sets.}

\item{background}{character; optional vector of at least 2 elements. \code{x}
will be filtered to only those elements of \code{background}.}

\item{min_size}{integer (\eqn{\geq 2}); the minimum set size.}

\item{max_size}{integer (\eqn{\geq 2}); the maximum set size.}
}
\value{
A named list of sets at most the same size as \code{x}.
}
\description{
Given a named list of sets, filter to those that contain at
  least \code{min_size} and no more than \code{min_size} elements. The sets
  are optionally filtered to elements of \code{background} before filtering
  by size.
}
\details{
Elements of \code{x} and \code{background} will be coerced to type
\code{character}.

If either \code{min_size} or \code{max_size} are less than 2, they will be
set to 2.
}
\examples{
x <- list("A" = c("a", "b", "c"),
          "B" = c("a", "a", "d", "e", NA), # duplicates and NA
          "C" = c("f", "g"))

# All sets have at least 2 elements, so this just removes duplicates and NA
filter_sets(x, min_size = 2L)

# Limit scope of sets before filtering
filter_sets(x, min_size = 2L, background = c("a", "c", "e", "z"))
}
\seealso{
\code{\link{find_aliases}}
}
