% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt_to_list.R
\name{gmt_to_list}
\alias{gmt_to_list}
\title{Convert GMT file to named list}
\usage{
gmt_to_list(path)
}
\arguments{
\item{path}{character; path to a GMT file.}
}
\value{
A named list of character vectors.
}
\description{
Create a named list of sets from a GMT file.
}
\note{
Equivalent to \code{fgsea::gmtPathways}.
}
\examples{
path <- system.file("extdata", "c5.go.v2023.2.Hs.symbols.gmt.gz",
                    package = "ostRich")

x <- gmt_to_list(path)

head(names(x)) # First 6 gene set names

x[1] # first set
}
