% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence}
\alias{incidence}
\title{Construct Sparse Incidence Matrix from List of Sets}
\usage{
incidence(x)
}
\arguments{
\item{x}{a named list of sets. Elements must be of type \code{"character"}.}
}
\value{
An object of class \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}
with unique set names as rows and unique elements as columns.
}
\description{
Given a named list of sets, construct a sparse incidence matrix.
}
\details{
The incidence matrix, \eqn{B}, is defined such that

\deqn{B_{ij} = \begin{cases} 1, & \text{if } \text{element } e_j \in
\text{set } s_i \\ 0, & \text{otherwise} \end{cases}}
}
\examples{
x <- list("A" = c("a", "b", "c"),
          "B" = c("c", "d"),
          "C" = c("x", "y", "z", "z"), # duplicates
          "D" = c("a", NA)) # missing values

(imat <- incidence(x))

# Sizes of sets and their pairwise intersections
tcrossprod(imat)

# Number of sets in which each element and pair of elements appears
crossprod(imat)

# Count number of elements unique to each set
keep <- apply(imat, 2, sum) == 1
apply(imat[, keep], 1, sum)
}
\seealso{
\code{\link{incidence_to_list}}, \code{\link{similarity}},
\code{\link[Matrix]{sparseMatrix}}
}
