% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_to_list.R
\name{incidence_to_list}
\alias{incidence_to_list}
\title{Convert incidence matrix to a named list of sets}
\usage{
incidence_to_list(incidence)
}
\arguments{
\item{incidence}{incidence matrix with set names as rows and elements as
columns. Usually, the output of \code{\link{incidence}}.}
}
\value{
a named list of sets with the same length as \code{nrow(incidence)}.
}
\description{
Converts an incidence matrix to a named list. The inverse of
  \code{\link{incidence}}.
}
\examples{
x <- list("A" = c("a", "b", "c"),
          "B" = c("c", "d"),
          "C" = c("x", "y", "z", "z"), # duplicates
          "D" = c("a", NA)) # missing values

(mat <- incidence(x)) # incidence matrix

incidence_to_list(mat)
}
