% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ostRich-package.R
\docType{package}
\name{ostRich-package}
\alias{ostRich}
\alias{ostRich-package}
\title{ostRich: Tools to Prepare Omics Sets for Enrichment Analysis/Gene Set Testing}
\description{
Tools to prepare lists of a priori biomolecular sets generated from high-throughput omics data, such as gene sets, for enrichment analysis/gene set testing.
}
\author{
\strong{Maintainer}: Tyler Sagendorf \email{tyler.sagendorf@pnnl.gov} (\href{https://orcid.org/0000-0003-1552-4870}{ORCID})

}
\keyword{internal}
