% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMSig-package.R
\docType{package}
\name{TMSig-package}
\alias{TMSig}
\alias{TMSig-package}
\title{TMSig: Tools for Molecular Signatures}
\description{
Tools to prepare and analyze lists of a priori molecular signatures, such as gene sets.
}
\author{
\strong{Maintainer}: Tyler Sagendorf \email{tyler.sagendorf@pnnl.gov} (\href{https://orcid.org/0000-0003-1552-4870}{ORCID})

Other contributors:
\itemize{
  \item Di Wu [contributor]
  \item Gordon Smyth [contributor]
}

}
\keyword{internal}
